+++ apps/vscode/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## 1.134.0 (Unreleased)
+
 ## 1.133.0 (Release on 2026-06-03)
 
 - Added diagnostics (i.e. squiggly underlines) to code cells in qmds (<https://github.com/quarto-dev/quarto/pull/980>).